refactor: 重构代码并优化依赖
- 更新 numpy、scipy、matplotlib 和 cvxpy 的最低版本要求
- 移除 Python 3.5 和 3.6 的支持，增加 Python 3.11 的支持
-优化部分代码以使用更高效的 numpy 数据类型
- 删除未使用的导入和警告
- 更新示例笔记本的内核名称
- 添加 install.bat 脚本

+++ doatools/estimation/source_number.py
@@ -89,7 +89,7 @@ def sorte(x):
     sigmas = [np.var(diffs[k:]) for k in range(n - 1)]
     s = np.zeros((n - 2,))
     for k in range(n - 2):
-        if np.abs(sigmas[k]) < np.finfo(np.float_).eps:
+        if np.abs(sigmas[k]) < np.finfo(np.float32).eps:
             s[k] = np.inf
         else:
             s[k] = sigmas[k + 1] / sigmas[k]
